% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome.R
\name{map_to_genome_x}
\alias{map_to_genome_x}
\title{Map positions to global x axis}
\usage{
map_to_genome_x(dt, genome_map, pos_col = "pos")
}
\arguments{
\item{dt}{A \code{data.table} with at least columns \code{"chrom"} and the
column named by \code{pos_col}.}

\item{genome_map}{A genome map from \code{\link{make_genome_map}}.}

\item{pos_col}{Name of the genomic position column. Default \code{"pos"}.}
}
\value{
The input \code{data.table} with \code{x_global} added (and the
  temporary \code{cum_start} column removed).
}
\description{
Given a \code{data.table} with \code{chrom} and a position
  column, adds a column \code{x_global} equal to the chromosome's cumulative
  start offset plus the local position.
}
\details{
Map genomic positions to cumulative x coordinates
}
\examples{
\dontrun{
dt[, pos := (start + end) \%/\% 2L]
dt <- map_to_genome_x(dt, gmap)
}

}
\seealso{
Other visualization: 
\code{\link{genome_from_bed}()},
\code{\link{make_genome_map}()},
\code{\link{plot_bed_igv}()},
\code{\link{plot_cluster_network_ggraph}()},
\code{\link{plot_genome_arrows_flat}()},
\code{\link{plot_genome_arrows_linear}()},
\code{\link{plot_node_reads_by_cluster}()},
\code{\link{plot_reads_around_node}()}
}
\concept{visualization}
