% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{plot_node_breaks_kde}
\alias{plot_node_breaks_kde}
\title{KDE plot of read breakpoints for a node}
\usage{
plot_node_breaks_kde(
  net,
  node,
  cluster_col = "chrom_cluster_id",
  bw_method = "SJ",
  bins = 50L,
  kernel = "gaussian",
  lw = 1
)
}
\arguments{
\item{net}{A network list from \code{\link{build_cluster_network}}.}

\item{node}{Character. The \code{chrom_cluster_id} of the node.}

\item{cluster_col}{Name of the cluster column in \code{net$reads}. Default
\code{"chrom_cluster_id"}.}

\item{bw_method}{Bandwidth selection method. Default \code{"SJ"}.}

\item{bins}{Integer. Number of histogram bins. Default \code{50L}.}

\item{kernel}{Kernel name. Default \code{"gaussian"}.}

\item{lw}{KDE line width. Default \code{1}.}
}
\value{
A named list with elements \code{plot} (a \code{ggplot2} object),
  \code{breaks} (numeric vector of start/end positions), \code{dens_df}
  (density data.table), \code{bw_kde}, and \code{bw_hist}.
}
\description{
For a chosen node, pools all read start and end positions and
  overlays a reflection-corrected KDE on top of a histogram to visualise
  breakpoint density.
}
\details{
Plot read breakpoint KDE for a network node
}
\examples{
\dontrun{
res <- plot_node_breaks_kde(net, node = "Chr1_12953910")
res$plot
}

}
\seealso{
Other KDE: 
\code{\link{estimate_modes_1d}()},
\code{\link{kde_reflect}()}
}
\concept{KDE}
